almost complete

+++ app/actions/message.action.ts
@@ -3,34 +3,31 @@ import { MessageSchema, MessageType } from "@/schemas/message.schema";
 import { connectDB } from "@/lib/db";
 import { z } from "zod";
 import Message from "@/models/Message";
-import { getAuthUser } from "@/lib/getAuthUser";
+import User from "@/models/User";
+import { headers } from "next/headers";
+import { auth } from "@/lib/auth";
 
-export interface SendMessageState {
-    success?: boolean;
-    message?: string;
-    error?: { content?: string[] };
-}
-
-export async function sendMessage(prevState: SendMessageState, formData: MessageType) {
+export async function sendMessage(formData: MessageType) {
     try {
         await connectDB();
 
         // validate the data against the schema
         const { success, data, error } = MessageSchema.safeParse(formData);
         if (!success) {
-            return { success: false, error: error.flatten().fieldErrors };
+            return { success: false, error: z.flattenError(error).fieldErrors };
         }
+
         // to whom we need to send the message
         const { content, to } = data;
-        // const user = await User.findOne({ username: to });
-        // if (!user) {
-        //     return { success: false, message: "User not found" };
-        // }
-        // if (!user.isAcceptingMessages) {
-        //     return { success: false, message: "User is not accepting messages" };
-        // }
+        const user = await User.findOne({ username: to }).select("isAcceptingMessages");
+        if (!user) {
+            return { success: false, message: "User not found" };
+        }
+        if (!user.isAcceptingMessages) {
+            return { success: false, message: "User is not accepting messages" };
+        }
 
-        // const message = await Message.create({ content, receiver: user._id });
+        const message = await Message.create({ content, receiver: user._id });
 
         await message.save();
 
@@ -43,14 +40,17 @@ export async function sendMessage(prevState: SendMessageState, formData: Message
 export async function deleteMessage({ messageId }: { messageId: string }) {
     try {
         await connectDB();
-        const loggedInUser = await getAuthUser();
-        if (!loggedInUser) {
+        const session = await auth.api.getSession({
+            headers: await headers(),
+        });
+        if (!session?.user) {
             return { success: false, message: "Unauthorized" };
         }
-        // const user = await User.findOne({ email: loggedInUser.email });
-        // if (!user) {
-        //     return { success: false, message: "User not found" };
-        // }
+        const user = await User.exists({ email: session?.user?.email });
+        if (!user) {
+            return { success: false, message: "Unauthorized" };
+        }
+
         // find the message
         const message = await Message.findById(messageId);
         if (!message) {
@@ -58,9 +58,9 @@ export async function deleteMessage({ messageId }: { messageId: string }) {
         }
 
         //   check if the message is sent to the user
-        // if (message.receiver.toString() !== user._id.toString()) {
-        //     return { success: false, message: "Unauthorized" };
-        // }
+        if (message.receiver.toString() !== user._id.toString()) {
+            return { success: false, message: "Unauthorized" };
+        }
 
         await message.deleteOne();
